refactor: separate local voice handlers and instructions
