+++ package.json
@@ -111,6 +111,9 @@
     },
     "./server/providers/regional": {
       "node": "./server/providers/regional.js"
+    },
+    "./server/providers/openai": {
+      "node": "./server/providers/openai.js"
     }
   }
 }
