+++ scripts/package-boundaries.json
@@ -294,5 +294,27 @@
       "src/data/regionalBrief.js"
     ],
     "external": []
+  },
+  "openai-provider": {
+    "runtime": "node",
+    "exports": [
+      "./server/providers/openai"
+    ],
+    "modules": [
+      "server/providers/common/rate-limit.js",
+      "server/providers/common/request.js",
+      "server/providers/common/source-root.js",
+      "server/providers/openai.js",
+      "server/providers/openai/constants.js",
+      "server/providers/openai/debug-log.js",
+      "server/providers/openai/hud-summary.js",
+      "server/providers/openai/instructions.js",
+      "server/providers/openai/rate-limit.js",
+      "server/providers/openai/realtime.js",
+      "server/providers/openai/tools.js",
+      "src/hudSummaryResponse.js",
+      "src/voice/voiceCost.js"
+    ],
+    "external": []
   }
 }
