+++ src/voice/gevActions.test.mjs
@@ -35,7 +35,7 @@ test('every live basemap is reachable by its own id — no enum value without a 
   assert.equal(normalizeStackId('Esri'), 'esri-imagery');
   assert.equal(normalizeStackId('esri imagery'), 'esri-imagery');
   // And the voice tool's enum must equal the set of live ids — no drift either way.
-  const config = readFileSync(new URL('../../server/providers/local.js', import.meta.url), 'utf8');
+  const config = readFileSync(new URL('../../server/providers/openai/tools.js', import.meta.url), 'utf8');
   const enumMatch = config.match(/enum: \[('photoreal'[^\]]*)\],\s*\n\s*description: 'photoreal = Google 3D/);
   assert.ok(enumMatch, 'set_map_stack enum literal must still be findable');
   const enumIds = enumMatch[1].split(',').map((s) => s.trim().replace(/^'|'$/g, ''));
